+++ src/components/modals/filters/handlers/AtomicFilterComponent.tsx
@@ -5,39 +5,20 @@ import { AtomicFilterComponentProps } from "cdm/ComponentsModel";
 import { OperatorFilter, StyleVariables } from "helpers/Constants";
 import React from "react";
 import ValueFilterComponent from "components/modals/filters/handlers/ValueFilterComponent";
-import FormControl from "@mui/material/FormControl";
-import Select, { SelectChangeEvent } from "@mui/material/Select";
-import { MenuItem } from "@mui/material";
-import modifyRecursiveFilterGroups from "./FiltersHelper";
-import OperatorSelectorComponent from "./OperatorSelectorComponent";
+import modifyRecursiveFilterGroups, {
+  ModifyFilterOptionsEnum,
+} from "components/modals/filters/handlers/FiltersHelper";
+import OperatorSelectorComponent from "components/modals/filters/handlers/OperatorSelectorComponent";
+import ExistedColumnSelectorComponent from "components/modals/filters/handlers/ExistedColumnSelectorComponent";
 
 const AtomicFilterComponent = (props: AtomicFilterComponentProps) => {
   const { table, recursiveIndex, level, atomicFilter, possibleColumns } = props;
   const { tableState } = table.options.meta;
   const { field, operator, value } = atomicFilter;
-  const [ddbbConfig, configActions] = tableState.configState((state) => [
-    state.ddbbConfig,
-    state.actions,
-  ]);
+  const configActions = tableState.configState((state) => state.actions);
   const configInfo = tableState.configState((state) => state.info);
   const columns = tableState.columns((state) => state.columns);
   const dataActions = tableState.data((state) => state.actions);
-
-  const onchangeExistedColumnHandler =
-    (conditionIndex: number[], level: number) =>
-    (event: SelectChangeEvent<string>, child: React.ReactNode) => {
-      const alteredFilterState = { ...configInfo.getFilters() };
-      modifyRecursiveFilterGroups(
-        possibleColumns,
-        alteredFilterState.conditions,
-        conditionIndex,
-        level,
-        "field",
-        event.target.value
-      );
-      configActions.alterFilters(alteredFilterState);
-      dataActions.dataviewRefresh(columns, ddbbConfig, alteredFilterState);
-    };
 
   const deleteConditionHadler =
     (conditionIndex: number[], level: number) => () => {
@@ -48,11 +29,15 @@ const AtomicFilterComponent = (props: AtomicFilterComponentProps) => {
         alteredFilterState.conditions,
         conditionIndex,
         level,
-        "delete"
+        ModifyFilterOptionsEnum.DELETE
       );
 
       configActions.alterFilters(alteredFilterState);
-      dataActions.dataviewRefresh(columns, ddbbConfig, alteredFilterState);
+      dataActions.dataviewRefresh(
+        columns,
+        configInfo.getLocalSettings(),
+        alteredFilterState
+      );
     };
 
   const onChangeFilterValueHandler =
@@ -64,56 +49,17 @@ const AtomicFilterComponent = (props: AtomicFilterComponentProps) => {
         alteredFilterState.conditions,
         conditionIndex,
         level,
-        "value",
+        ModifyFilterOptionsEnum.VALUE,
         event.target.value
       );
       configActions.alterFilters(alteredFilterState);
-      dataActions.dataviewRefresh(columns, ddbbConfig, alteredFilterState);
+      dataActions.dataviewRefresh(
+        columns,
+        configInfo.getLocalSettings(),
+        alteredFilterState
+      );
     };
 
-  const existedColumnSelector = (selectorProps: {
-    currentCol: string;
-    recursiveIndex: number[];
-    level: number;
-  }) => {
-    const { currentCol, recursiveIndex, level } = selectorProps;
-    return (
-      <FormControl
-        fullWidth
-        key={`FormControl-existedColumnSelector-${level}-${recursiveIndex[level]}`}
-      >
-        <Select
-          value={currentCol}
-          size="small"
-          key={`Select-existedColumnSelector-${level}-${recursiveIndex[level]}`}
-          onChange={onchangeExistedColumnHandler(recursiveIndex, level)}
-          style={{
-            backgroundColor: StyleVariables.BACKGROUND_PRIMARY,
-            color: StyleVariables.TEXT_NORMAL,
-          }}
-          MenuProps={{
-            PaperProps: {
-              sx: {
-                backgroundColor: StyleVariables.BACKGROUND_PRIMARY,
-                color: StyleVariables.TEXT_NORMAL,
-              },
-            },
-          }}
-        >
-          {possibleColumns.map((key) => {
-            return (
-              <MenuItem
-                value={key}
-                key={`MenuItem-existedColumnSelector-${key}--${level}-${recursiveIndex[level]}}`}
-              >
-                {key}
-              </MenuItem>
-            );
-          })}
-        </Select>
-      </FormControl>
-    );
-  };
   return (
     <Grid
       container
@@ -126,23 +72,25 @@ const AtomicFilterComponent = (props: AtomicFilterComponentProps) => {
         xs="auto"
         key={`Grid-AtomicFilter-field-${level}-${recursiveIndex[level]}`}
       >
-        {existedColumnSelector({
-          currentCol: field,
-          recursiveIndex: recursiveIndex,
-          level: level,
-        })}
+        <ExistedColumnSelectorComponent
+          currentCol={field}
+          recursiveIndex={recursiveIndex}
+          level={level}
+          table={table}
+          possibleColumns={possibleColumns}
+        />
       </Grid>
       <Grid
         item
         xs="auto"
         key={`Grid-AtomicFilter-operator-${level}-${recursiveIndex[level]}`}
       >
-        {OperatorSelectorComponent({
-          currentOp: operator,
-          recursiveIndex: recursiveIndex,
-          level: level,
-          table: table,
-        })}
+        <OperatorSelectorComponent
+          currentOp={operator}
+          recursiveIndex={recursiveIndex}
+          level={level}
+          table={table}
+        />
       </Grid>
       {/* if value exists, show it */}
       {![OperatorFilter.IS_EMPTY[0], OperatorFilter.IS_NOT_EMPTY[0]].contains(
